project/requirements.txt
Django==1.10.6
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-haystack==2.6.1
djangorestframework==3.6.4
Markdown==2.6.9
Pillow==4.3.0
